dynamically construct the paths to the training and testing data instead of hardcoding them.
To make the code more portable and robust, I dynamically constructed the paths to the training and testing data instead of hardcoding them. By assuming the archive (sport Balls) folder is in the same directory as the Python script. I did the same for the jupyter notebook but by assuming that  the Project folder is a sibling of the Notebooks folder.