Adding PCA
PCA was able to improve the accuracy at about 50%, this feature extraction is also much faster to execute which allowed me to add more parameters to the GridSearchCV hence a more accurate hyperparameter optimization.
Some minor adjustments to the code such as adding specific prints at every key point of the code to keep track of where in the code the kernel is executing, removing a useless parts of the code, and condensing some lines.